Simplify API, there is little to be discovered without the start directory being importable (except when the discovered test are not python code). So let the calle setup sys.path as he sees fit instead of trying to guess. From there, there is no point passing around the top level dir nor the current dir.
Add DirLoader.discover_paths to better control which paths are tried.

Default to PackageLoader and ModuleLoader for TestLoader, that's the
intended use case (tests written in python, alongside the code they're
testing and as such importable from there giving a proper sys.path). The
rationale is that there is little point allowing tests to be imported
with a sys.path different than the one used to import the code.

Start implementing tests for DirLoader and PackageLoader, defining some
more tests along the way.