Don't rely on shebang lines working in the virtualenv.
The ci.sh script falls down when the full path to the working directory
has spaces in it (which is common for Jenkins workspaces).  This is
because it's not legal to have spaces in the interpreter part of a
shebang line, and scripts like pip and flake8 are used from the
virtualenv.

(virtualenv itself used to have this problem too, as it called
easy_install in the newly created env to set up pip, but this is fixed in
later versions)

In order to fix this, the paths to the pip and flake8 scripts are passed
to the virtualenv python